chore(example): convert App.js to App.tsx to match new react-native
