+++ refresh-example.sh
@@ -16,7 +16,7 @@ else
   mkdir -p TEMP/ios/RNFBSDKExample
   cp RNFBSDKExample/README.md TEMP/
   cp RNFBSDKExample/android/local.properties TEMP/android/ || true
-  cp RNFBSDKExample/App.js TEMP/
+  cp RNFBSDKExample/App.tsx TEMP/
 fi
 
 # Purge the old sample
