+++ tsconfig.json
@@ -1,4 +1,7 @@
 {
+  "exclude": [
+    "RNFBSDKExample"
+  ],
   "compilerOptions": {
     /* Visit https://aka.ms/tsconfig.json to read more about this file */
